+++ src/collective/sidebar/setuphandlers.py
@@ -9,7 +9,11 @@ class HiddenProfiles(object):
 
     def getNonInstallableProfiles(self):
         """Hide uninstall profile from site-creation and quickinstaller"""
-        return ['collective.sidebar:uninstall']
+        return [
+            'collective.sidebar:uninstall',
+            'collective.sidebar:to_1001',
+            'collective.sidebar:to_1300',
+        ]
 
 
 def post_install(context):