+++ src/collective/sidebar/upgrades.py
@@ -18,9 +18,13 @@ def reload_gs_profile(context):
 def to_1001(context):
     context.runAllImportStepsFromProfile('profile-collective.sidebar:to_1001')
     logger.info('Removed collective.sidebar Resources')
-    p = api.portal
-    p.set_registry_record(name='collective.sidebar.root_nav', value=False)  # noqa
-    p.set_registry_record(name='collective.sidebar.enable_actions', value=True)  # noqa
-    p.set_registry_record(name='collective.sidebar.enable_collapse', value=False)  # noqa
-    p.set_registry_record(name='collective.sidebar.enable_cookies', value=False)  # noqa
+    api.portal.set_registry_record(name='collective.sidebar.root_nav', value=False)  # noqa
+    api.portal.set_registry_record(name='collective.sidebar.enable_actions', value=True)  # noqa
+    api.portal.set_registry_record(name='collective.sidebar.enable_collapse', value=False)  # noqa
+    api.portal.set_registry_record(name='collective.sidebar.enable_cookies', value=False)  # noqa
+    logger.info('Added collective.sidebar registry records')
+
+
+def to_1300(context):
+    context.runAllImportStepsFromProfile('profile-collective.sidebar:to_1300')
     logger.info('Added collective.sidebar registry records')